### 2025-08-14 #8808 @editorialbot
[PRE REVIEW]: JumpMetrics: A Python package computing countermovement and squat jump events and metrics
**Submitting author:** @stevenhirsch (<a href="http://orcid.org/0000-0002-4394-1922">Steven Mark Hirsch</a>)
**Repository:** https://github.com/stevenhirsch/force-plate-jump-analyses
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
