### 2025-08-15 #8808 @Kevin-Mattheus-Moerman
- [ ] Recommendation: I recommend that you add some visualisations around the software functionality (e.g. graphs) to the README so new users quickly get insight into your project's capabilities.
