### 2025-09-01 ee69819
added DOI to references, added a contributing.md file, and updated the readme
