+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={96--106},
   year={2018},
-  publisher={LWW}
+  publisher={LWW},
+  doi={10.1519/SSC.0000000000000375}
 }
 
 @article{owen:2014,
@@ -17,7 +18,8 @@
   number={6},
   pages={1552--1558},
   year={2014},
-  publisher={LWW}
+  publisher={LWW},
+  doi={10.1519/JSC.0000000000000311}
 }
 
 @article{mcmahon:2017,
@@ -28,7 +30,8 @@
   number={6},
   pages={803--811},
   year={2017},
-  publisher={Human Kinetics, Inc.}
+  publisher={Human Kinetics, Inc.},
+  doi={10.1123/ijspp.2016-0467}
 }
 
 @article{bird:2022,
@@ -38,7 +41,8 @@
   volume={13},
   pages={868002},
   year={2022},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={10.3389/fphys.2022.868002}
 }
 
 @article{pedley2020,
@@ -48,7 +52,8 @@
   volume={2},
   pages={287--304},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s42978-020-00090-1}
 }
 
 @article{vanHooren2017,
@@ -59,7 +64,8 @@
   number={7},
   pages={2011--2020},
   year={2017},
-  publisher={LWW}
+  publisher={LWW},
+  doi={10.1519/JSC.0000000000001913}
 }
 
 @article{watkins:2017,
@@ -70,7 +76,8 @@
   number={12},
   pages={3305--3310},
   year={2017},
-  publisher={LWW}
+  publisher={LWW},
+  doi={10.1519/JSC.0000000000002231}
 }
 
 @article{xu:2023,
@@ -81,5 +88,6 @@
   number={5},
   pages={1055--1072},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s40279-023-01828-x}
 }
