+++ docs/development/contributing.md
@@ -1,0 +1,248 @@
+# Contributing to JumpMetrics
+
+Thank you for your interest in contributing to JumpMetrics! 🎉  
+We welcome contributions from researchers, practitioners, students, and developers of all backgrounds. Whether you're fixing a typo, improving documentation, or adding a new biomechanical metric, your contribution is valued.
+
+This document provides guidelines and instructions to help you get started.
+
+## Table of Contents
+- [Code of Conduct](#code-of-conduct)
+- [Getting Started](#getting-started)
+- [Development Setup](#development-setup)
+- [Project Structure](#project-structure)
+- [Contribution Scope](#contribution-scope)
+- [Making Contributions](#making-contributions)
+- [Pull Request Process](#pull-request-process)
+- [Coding Standards](#coding-standards)
+- [Testing Guidelines](#testing-guidelines)
+- [Documentation](#documentation)
+- [Reporting Issues](#reporting-issues)
+
+## Code of Conduct
+
+By participating in this project, you agree to maintain a respectful and inclusive environment. We expect all contributors to:
+- Use welcoming and inclusive language
+- Be respectful of differing viewpoints and experiences
+- Gracefully accept constructive criticism
+- Focus on what is best for the community
+- Show empathy towards other community members
+- Respect data sharing ethics and privacy considerations when working with human subject data
+- Provide proper attribution for contributed algorithms and methods
+- Foster collaborative relationships with researchers and practitioners in biomechanics
+
+## Getting Started
+
+1. Fork the repository on GitHub
+2. Clone your fork locally:
+```bash
+git clone https://github.com/stevenhirsch/force-plate-jump-analyses.git
+cd force-plate-jump-analyses
+```
+3. Add the upstream repository:
+```bash
+git remote add upstream https://github.com/stevenhirsch/force-plate-jump-analyses.git
+```
+
+## Development Setup
+
+1. Create and activate a conda environment using the provided environment files:
+```bash
+conda env create -f env.yml
+conda activate jumpmetrics
+conda env update -f local_env.yml
+```
+
+`env.yml` contains the minimal packages to build and run the package. `local_env.yml` contains additional packages for local development and batch processing.
+
+2. Install the package in development mode:
+```bash
+pip install -e .
+```
+
+### Pre-commit Setup (Optional but Recommended)
+
+To maintain code quality and consistency, we recommend setting up pre-commit hooks:
+
+```bash
+pip install pre-commit
+pre-commit install
+```
+
+This will automatically format code and run basic checks before commits.
+
+## Project Structure
+
+Understanding the project layout will help you navigate and contribute effectively:
+
+```
+force-plate-jump-analyses/
+├── jumpmetrics/           # Main package code
+│   ├── core/             # Core analysis functions
+│   ├── events/           # Event detection
+│   ├── metrics/          # Jump metric calculations
+│   └── signal_processing/ # Signal processing and filtering
+├── tests/                # Test suite
+│   ├── core/            
+│   ├── signal_processing/
+│   └── example_data/    # Test data files
+├── docs/                 # Documentation
+├── analyses/            # Study-specific analyses
+├── docker_example/      # Docker usage example
+├── paper/               # Scientific paper
+└── env.yml             # Environment specification
+```
+
+## Contribution Scope
+
+There are many ways you can contribute, including:
+- Fixing bugs or improving code readability
+- Adding new jump metrics or biomechanical analyses
+- Improving event detection or signal processing methods
+- Writing or expanding documentation and tutorials
+- Adding or refining tests to improve coverage and reliability
+- Improving performance for large datasets
+- Helping improve reproducibility (Docker, CI, etc.)
+
+Both small and large contributions are welcome!
+
+## Making Contributions
+
[…]
