### 2025-09-01 #8808 @stevenhirsch
> [@stevenhirsch](https://github.com/stevenhirsch) Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:
>
>     * [x]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.[x]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).[x]  Recommendation: I recommend that you add some visualisations around the software functionality (e.g. graphs) to the README so new users quickly get insight into your project's capabilities.

@Kevin-Mattheus-Moerman Thanks for the feedback! I've addressed these comments and the DOI issues are now resolved.

Also re: the above message:
> ⚠️ Important note to authors: Please do not directly contact the authors of these papers. If you believe these papers are relevant to your submission, please mention this in your submission or communicate with your handling editor.

The papers are all relevant to my submission if that helps at all with the speeding up the waitlist process! In the meantime, please let me know if you need anything else from me.
