### 2025-09-15 6427ed0
updated repo to use pixi
