+++ .github/workflows/python-tests.yml
@@ -1,37 +1,34 @@
-name: Python workflow
+name: Python Tests
 
-on: [push]
+on: [push, pull_request]
 
 jobs:
-  python_ci:
-    runs-on: ubuntu-latest
+  test:
+    name: Test on ${{ matrix.os }} with Python ${{ matrix.python-version }}
+    runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.10"]
+        os: [ubuntu-latest, windows-latest, macos-latest, macos-14]
+        python-version: ["3.10", "3.11", "3.12"]
+        exclude:
+          - os: macos-latest
+            python-version: "3.11"
 
     steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Add conda to system path
-      run: |
-        # $CONDA is an environment variable pointing to the root of the miniconda directory
-        echo $CONDA/bin >> $GITHUB_PATH
-    - name: Explictly install python 3.10 with Conda
-      run: |
-        conda install python=3.10
-    - name: Install dependencies
-      run: |
-        conda env update --file env.yml --name base
-    - name: Test with pytest
-      run: |
-        conda install pytest pytest-cov
-        pytest ./tests --cov --cov-report=xml
+    - uses: actions/checkout@v4
 
-    - name: Check types with mypy
-      run: |
-        conda install mypy 
-        mypy ./jumpmetrics --ignore-missing-imports --follow-imports=skip --verbose
+    - name: Install Pixi
+      uses: prefix-dev/setup-pixi@v0.8.1
+      with:
+        pixi-version: v0.54.2
+        cache: true
+
+    - name: Install dependencies
+      run: pixi install
+
+    - name: Run tests
+      run: pixi run -e test test
+
+    - name: Run type checking
+      run: pixi run -e test typecheck
