### 2025-09-17 a70ea0b
Modernize build system and update documentation
- Remove outdated build files (setup.py, env.yml, local_env.yml)
- Update README with improved pixi-based installation instructions
- Fix pyproject.toml task paths for batch processing scripts
- Add CLAUDE.md for future development guidance
- Standardize on pixi for environment and dependency management
