+++ local_env.yml
@@ -1,14 +1,0 @@
-channels:
-  - defaults
-dependencies:
-  - ipython=8.12.*
-  - jupyter=1.0.*
-  - jupyterlab=3.6.*
-  - tqdm
-  - pip
-  - build
-  - pytest
-  - mypy
-  - pip:
-    - flake8
-    - build
