+++ env.yml
@@ -1,11 +1,0 @@
-name: jumpmetrics
-channels:
-  - defaults
-  - conda-forge
-dependencies:
-  - python>=3.10
-  - pandas=2.0.*
-  - matplotlib
-  - scipy=1.14.*
-  - scikit-learn=1.3.*
-  - pip=23.*
