### 2025-09-20 e541d66
added additional testing to improve coverage of all functions
