+++ tests/signal_processing/test_numerical.py
@@ -74,3 +74,200 @@ def test_compute_integral_of_signal_with_pandas():
 
     expected = -(1 / (2 * np.pi)) * np.cos(2 * np.pi * t) + (1 / (2 * np.pi))
     assert np.allclose(integral, expected, atol=1e-3)
+
+
+def test_integrate_area_edge_cases():
+    """Test edge cases for integrate_area function"""
+
+    # Test with constant signal
+    time = np.linspace(0, 1, 100)
+    signal = np.full(100, 5.0)
+
+    area_trap = integrate_area(time, signal, method='trapezoidal')
+    area_simp = integrate_area(time, signal, method='simpsons')
+
+    # Area should be width * height = 1 * 5 = 5
+    assert np.isclose(area_trap, 5.0, rtol=1e-3)
+    assert np.isclose(area_simp, 5.0, rtol=1e-3)
+
+    # Test with negative values
+    signal_neg = np.full(100, -3.0)
+    area_neg = integrate_area(time, signal_neg, method='trapezoidal')
+    assert np.isclose(area_neg, -3.0, rtol=1e-3)
+
+    # Test with single point (should return 0)
+    time_single = np.array([0])
+    signal_single = np.array([5])
+    area_single = integrate_area(time_single, signal_single, method='trapezoidal')
+    assert area_single == 0
+
+    # Test with two points
+    time_two = np.array([0, 1])
+    signal_two = np.array([0, 2])
+    area_two = integrate_area(time_two, signal_two, method='trapezoidal')
+    assert np.isclose(area_two, 1.0)  # Triangle area = 0.5 * base * height
+
+
+def test_get_finite_difference_coefficients_edge_cases():
+    """Test edge cases for finite difference coefficients"""
+
+    # ACTUAL BEHAVIOR: Function requires even order_accuracy for symmetric difference
+    # Test that odd order accuracy raises appropriate error
+    with pytest.raises(ValueError, match="Order of Accuracy must be a multiple of 2"):
+        get_finite_difference_coefficients(1, 1)  # Odd order not allowed
+
+    with pytest.raises(ValueError, match="Order of Accuracy must be a multiple of 2"):
+        get_finite_difference_coefficients(1, 3)  # Odd order not allowed
+
+    # Test valid even order accuracy
+    coeff_1st_2nd = get_finite_difference_coefficients(1, 2)
+    assert len(coeff_1st_2nd) == 3  # Should have 3 coefficients
+
+    # Test higher derivatives with appropriate even accuracy
+    coeff_3rd_4th = get_finite_difference_coefficients(3, 4)
+    assert len(coeff_3rd_4th) == 7  # Should have 7 coefficients for 3rd derivative, 4th order
+
+    # Test second derivative with even order
+    coeff_2nd_2nd = get_finite_difference_coefficients(2, 2)
+    assert len(coeff_2nd_2nd) == 3  # Should have 3 coefficients
+
+
+def test_compute_derivative_edge_cases():
+    """Test edge cases for compute_derivative function"""
+
+    # Test with constant signal (derivative should be zero)
+    y_constant = np.full(10, 5.0)
+    t = 0.1
+    dy_dt = compute_derivative(y_constant, t, 1)
+
+    # Interior points should be approximately zero
+    assert np.allclose(dy_dt[1:-1], 0, atol=1e-10)
+    # Boundary points should be NaN
+    assert np.isnan(dy_dt[0]) and np.isnan(dy_dt[-1])
+
+    # Test with linear signal (derivative should be constant)
+    y_linear = np.linspace(0, 10, 11)  # Slope = 1
+    dy_dt_linear = compute_derivative(y_linear, 1.0, 1)
+
[…]
