### 2025-09-20 9f0767f
handling edge cases of very short data gracefully and updated tests
