### 2025-09-24 09ca6c3
removed outdated analyses results and scripts. Updated README
