+++ study_1_batch_process.py
@@ -1,198 +1,0 @@
-"""Batch processing script for study 1"""
-import os
-import pandas as pd
-import numpy as np
-from tqdm import tqdm
-import matplotlib.pyplot as plt
-from jumpmetrics.core.processors import ForceTimeCurveCMJTakeoffProcessor
-from jumpmetrics.core.io import load_cropped_force_data
-from jumpmetrics.signal_processing.filters import butterworth_filter
-
-
-CREATE_PLOTS = True
-main_dir = os.path.join(os.getcwd(), 'analyses', 'study_1')
-data_dir = os.path.join(main_dir, 'raw_data')
-fig_dir = os.path.join(main_dir, 'figures')
-kinematic_data_dir = os.path.join(main_dir, 'kinematic_data')
-all_filenames = os.listdir(data_dir)
-xlsx_data_filepath = os.path.join(main_dir, 'signal_conditioning_filter_cutoffs.xlsx')
-xlsx_data = pd.read_excel(xlsx_data_filepath)
-full_dataset = pd.DataFrame()
-
-for filename in tqdm(all_filenames):
-    filepath = os.path.join(data_dir, filename)
-    FILE_PREFIX = '_'.join(filename.split('_')[0:2])
-    pid = FILE_PREFIX.split('_', maxsplit=1)[0]
-    trial_num = FILE_PREFIX.split('_')[1]
-    pid_fig_dir = os.path.join(fig_dir, pid, trial_num)
-    pid_data_dir = os.path.join(kinematic_data_dir, pid, trial_num)
-    if not os.path.exists(pid_fig_dir):
-        os.makedirs(pid_fig_dir)
-    if not os.path.exists(pid_data_dir):
-        os.makedirs(pid_data_dir)
-    participant_filters = xlsx_data[xlsx_data['file_prefix'] == FILE_PREFIX]
-    for col in participant_filters.columns:
-        if col == 'file_prefix':
-            continue
-        else:
-            try:
-                tmp_df = pd.DataFrame()
-                tmp_df['file_prefix'] = [FILE_PREFIX]
-                tmp_df['cutoff_type'] = [col]
-                cutoff_frequency = participant_filters[col].item()
-                pid_fig_dir_cutoff_type = os.path.join(pid_fig_dir, col)
-                if not os.path.exists(pid_fig_dir_cutoff_type):
-                    os.makedirs(pid_fig_dir_cutoff_type)
-                tmp_df['cutoff_frequency'] = [cutoff_frequency]
-                force_series = load_cropped_force_data(
-                    filepath=filepath,
-                    freq=None
-                )
-                if cutoff_frequency != 'none':
-                    filtered_force_series = butterworth_filter(
-                        arr=force_series,
-                        cutoff_frequency=cutoff_frequency,
-                        fps=2000,
-                        padding=2000
-                    )
-                    CMJ = ForceTimeCurveCMJTakeoffProcessor(
-                        force_series=filtered_force_series[-4000:],
-                        sampling_frequency=2000
-                    )
-                else:
-                    # process the raw data
-                    CMJ = ForceTimeCurveCMJTakeoffProcessor(
-                        force_series=np.array(force_series)[-4000:],
-                        sampling_frequency=2000
-                    )
-                CMJ.get_jump_events()
-                CMJ.compute_jump_metrics()
-                CMJ.create_jump_metrics_dataframe(pid=pid)
-                CMJ.create_kinematic_dataframe()
-                if CREATE_PLOTS:
-                    CMJ.plot_waveform(
-                        waveform_type='force',
-                        title=filename + '_' + col,
-                        savefig=True,
-                        figname=os.path.join(pid_fig_dir_cutoff_type, 'force.png')
-                    )
-                    CMJ.plot_waveform(
[…]
