+++ study_3_batch_process.py
@@ -1,312 +1,0 @@
-"""Batch processing script for study 3"""
-import os
-import logging
-import pandas as pd
-from tqdm import tqdm
-import matplotlib.pyplot as plt
-from jumpmetrics.core.processors import (
-    ForceTimeCurveCMJTakeoffProcessor, ForceTimeCurveSQJTakeoffProcessor
-)
-from jumpmetrics.core.io import (
-    load_raw_force_data_with_no_column_headers, sum_dual_force_components,
-    find_first_frame_where_force_exceeds_threshold,
-    find_frame_when_off_plate, get_n_seconds_before_takeoff
-)
-from jumpmetrics.signal_processing.filters import butterworth_filter
-
-class WarningCatcher(logging.Handler):
-    """Class for catching warnings and saving them"""
-    def __init__(self):
-        super().__init__()
-        self.warnings = []
-
-    def emit(self, record):
-        if record.levelno == logging.WARNING:
-            self.warnings.append(record.getMessage())
-
-    def reset(self):
-        """Reset warnings"""
-        self.warnings = []
-
-# Set up the warning catcher
-warning_catcher = WarningCatcher()
-logging.getLogger().addHandler(warning_catcher)
-
-# Parsing funcs
-def parse_jump_info(jump_info):
-    """Function to parse jump info"""
-    jump_type = jump_info[:3]
-    file_code = jump_info[3:].split('.txt')[0]
-    overall_recording_num = file_code[-2:]
-    rest_of_file_code = file_code[:-2]
-    num_zeros = rest_of_file_code.count('0')
-    trial_num_in_block = rest_of_file_code[0] if num_zeros == 3 else rest_of_file_code[0:2]
-    return jump_type, overall_recording_num, trial_num_in_block
-
-
-def parse_filename_metadata(fname):
-    """Function to parse filename metadata"""
-    sport, pid, jump_info = fname.split('_')
-    jump_type, overall_recording_num, trial_num_in_block = parse_jump_info(
-        jump_info=jump_info
-    )
-    filename_metadata = {
-        'sport': sport,
-        'pid': pid,
-        'jump_type': jump_type,
-        'trial_num_in_block': int(trial_num_in_block),
-        'overall_recording_num': int(overall_recording_num)
-
-    }
-    return filename_metadata
-
-CREATE_PLOTS = False
-main_dir = os.path.join(os.getcwd(), 'analyses', 'study_3')
-data_dir = os.path.join(main_dir, 'raw_data')
-fig_dir = os.path.join(main_dir, 'figures')
-kinematic_data_dir = os.path.join(main_dir, 'kinematic_data')
-all_filenames = os.listdir(data_dir)
-# xlsx_data_filepath = os.path.join(main_dir, 'signal_conditioning_filter_cutoffs.xlsx')
-# xlsx_data = pd.read_excel(xlsx_data_filepath)
-full_cmj_dataset = pd.DataFrame()
-full_sqj_dataset = pd.DataFrame()
-# xlsx_data['pid'] = xlsx_data['file_prefix'].str.split('_').str[0]
-# xlsx_data_condensed = xlsx_data.groupby('pid').mean(numeric_only=True)
-# xlsx_data_condensed.drop(['trial_cutoff', 'participant_cutoff', 'literature_cutoff'], axis=1, inplace=True)
-# xlsx_data_condensed.reset_index(inplace=True)
-corrupt_filenames = [
-    'FHOC_P27_CMJ100002.txt',
-    'FHOC_P12_SQT200006.txt'
-]
-
-for filename in tqdm(all_filenames):
-    if filename in corrupt_filenames:
-        continue
-    warning_catcher.reset()
-    filepath = os.path.join(data_dir, filename)
[…]

[…]
