### 2025-09-24 5fb249c
updated images for paper
