+++ paper/paper.md
@@ -35,15 +35,15 @@ Researchers and practitioners (e.g., sports team scientists, analysts, or coache
 ## Event Detections and Metrics
 `JumpMetrics` computes various events and metrics for the countermovement and squat jump leveraging the vertical axis data from a force plate. The events (and thus metrics) are slightly different between the jump variations, given the differences in their movement executions. There are classes for processing various phases of the countermovement jump, a helper function to process the entire jump and landing, as well as individual functions for even more granularity for analyses. Furthermore, `JumpMetrics` computes the vertical axis acceleration, velocity, and displacement of the estimated center of mass trajectory for each frame of data, irrespective of whether one is using a triaxial or uniaxial force plate. These data are computed by first dividing the force trace by the individual's computed bodymass to obtain the acceleration data. Then, the acceleration signal is integrated to compute the instantaneous velocity. Finally, the signal is integrated one more time to compute the instataneous displacement. Some examples of these computed waveforms are shown in Figures 1, 2, and 3.
 
-![Example acceleration waveform (i.e., the raw force trace divided by the body mass, which was estimated from bodyweight).](../analyses/study_2/figures/F02/CTRL1/group_cutoff/acceleration.png){ width=100% }
+![Example acceleration waveform (i.e., the raw force trace divided by the body mass, which was estimated from bodyweight).](../example_images/acceleration.png){ width=100% }
 
 Figure 1. Example countermovement jump acceleration trace with events detected during the takeoff phase. Positive accelerations represent the center of mass accelerating upwards, whereas negative accelerations represent the center of mass accelerating downwards towards the floor/force plate.
 
-![Example velocity waveform (i.e., the integrated acceleration waveform, assuming 0 velocity during quiet standing).](../analyses/study_2/figures/F02/CTRL1/group_cutoff/velocity.png){ width=100% }
+![Example velocity waveform (i.e., the integrated acceleration waveform, assuming 0 velocity during quiet standing).](../example_images/velocity.png){ width=100% }
 
 Figure 2. Example countermovement jump velocity trace with events detected during the takeoff phase. Positive velocities represent the center of mass is moving upwards, whereas negative velocities represent the center of mass moving downwards.
 
-![Example displacement waveform (i.e., the integrated velocity waveform, assuming 0 displacement during quiet standing).](../analyses/study_2/figures/F02/CTRL1/group_cutoff/displacement.png){ width=100% }
+![Example displacement waveform (i.e., the integrated velocity waveform, assuming 0 displacement during quiet standing).](../example_images/displacement.png){ width=100% }
 
 Figure 3. Example countermovement jump displacement trace with events detected during the takeoff phase. Positive displacements represent the center of mass being higher relative to quiet standing, whereas negative displacements represent the center of mass being closer to the floor relative to quiet standing.
 
@@ -55,7 +55,7 @@ There are two main phases preceeding the moment of takeoff during countermovemen
 
[…]
