+++ pyproject.toml
@@ -3,7 +3,7 @@ authors = [{name = "stevenhirsch", email = "stevehirsch94@gmail.com"}]
 dependencies = []
 name = "JumpMetrics"
 requires-python = ">= 3.10"
-version = "0.1.0"
+version = "0.1.1"
 
 [build-system]
 build-backend = "hatchling.build"
