### 2026-02-14 78e100f
updated paper to conform to updated JOSS requirements
