+++ paper/paper.bib
@@ -91,3 +91,14 @@
   publisher={Springer},
   doi={10.1007/s40279-023-01828-x}
 }
+
+@article{dixon:2017,
+  title={biomechZoo: An open-source toolbox for the processing, analysis, and visualization of biomechanical movement data},
+  author={Dixon, Philippe C and Sch{\"u}tte, Kevin H and Vanwanseele, Benedicte and Jacobs, Jennifer V and Dennerlein, Jack T and Schiffman, Jordyn M},
+  journal={Computer Methods and Programs in Biomedicine},
+  volume={140},
+  pages={1--10},
+  year={2017},
+  publisher={Elsevier},
+  doi={10.1016/j.cmpb.2016.11.007}
+}
