### 2026-02-14 #8808 @stevenhirsch
@arfon Thank you for the update. I have made the necessary changes to the paper to conform to the updated journal requirements. I hope you will find that it still remains in the scope of the journal. Please let me know if you have any other questions!
