### 2026-02-15 #8808 @arfon
@stevenhirsch – in your research impact statement you say _"JumpMetrics has begun establishing itself within the biomechanics research and sports science communities. The software is currently being used by external researchers and practitioners for force plate jump analysis"_

Can you provide direct evidence of this?
