### 2026-02-21 #8808 @stevenhirsch
> [@stevenhirsch](https://github.com/stevenhirsch) – in your research impact statement you say _"JumpMetrics has begun establishing itself within the biomechanics research and sports science communities. The software is currently being used by external researchers and practitioners for force plate jump analysis"_
>
> Can you provide direct evidence of this?

@arfon I can provide anecdotal evidence of colleagues and others (both in academia and in other professional settings) using, or planning to use, this software for various purposes. I believe that I can provide more direct evidence once this becomes easier to cite!
