### 2026-03-16 #8808 @Kevin-Mattheus-Moerman
@stevenhirsch if the other works (hinted at in your discussion with Arfon ☝️) using your library have preprints already, you can cite them in this paper if possible.
