### 2026-05-21 bad957d
updated repo structure, package dependencies, and paper for JOSS
