+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Summary
 
-Researchers and practitioners (e.g., sports team scientists, analysts, or coaches) commonly assess countermovement jump and squat jump performance on force plates. A countermovement jump involves a vertical jump where the jumper first dips downwards before immediately jumping upwards as high as possible. In contrast, a squat jump is a vertical jump whereby the jumper pauses briefly in the bottom "squat" position after dipping downards, and then jumps upwards to minimize contributions from the stretch-shortening cycle. A scientist, analyst, or coach can use data from these vertical jump variations for various reasons, such as to evaluate people's neuromuscular capacity, injury risk, or readiness/preparedness for high-intensity training. For evaluating capacity, some researchers have examined variables such as maximum jump height, peak force, rate of force development, and impulse `[@mcmahon:2017]`. When assessing injury risk, researchers have examined landing forces (e.g., `[@pedley2020]`) or have leveraged statistical techniques on various countermovement jump variables simultaneously `[@bird:2022]`. For training readiness, researchers have measured changes in vertical jump height between consecutive training sessions (e.g., `[@watkins:2017]`). Examining the difference in performance between the two jump variations may also provide insights into the strengths and weaknesses of the athlete to direct future training `[@vanHooren2017]`. For example, some researchers may compute an "eccentric utilization ratio" by comparing metrics from the countermovement jump relative to the squat jump and use that to inform whether someone should focus their training on improving their ability to leverage the stretch shortening cycle to maximize jumping performance `[@vanHooren2017]`. Although it is common to collect this kinetic data from a force plate for these jump variations for several applications, there are currently no free, open-source resources to detect events and compute metrics for reproducible and accessible data processing. `JumpMetrics` fills this gap in both applied practice and in the sports science literature.
+Researchers and practitioners (e.g., sports team scientists, analysts, or coaches) commonly assess countermovement jump and squat jump performance on force plates. A countermovement jump involves a vertical jump where the jumper first dips downwards before immediately jumping upwards as high as possible. In contrast, a squat jump is a vertical jump whereby the jumper pauses briefly in the bottom "squat" position after dipping downards, and then jumps upwards to minimize contributions from the stretch-shortening cycle. A scientist, analyst, or coach can use data from these vertical jump variations for various reasons, such as to evaluate people's neuromuscular capacity, injury risk, or readiness/preparedness for high-intensity training. For evaluating capacity, some researchers have examined variables such as maximum jump height, peak force, rate of force development, and impulse [@mcmahon:2017]. When assessing injury risk, researchers have examined landing forces (e.g., [@pedley2020]) or have leveraged statistical techniques on various countermovement jump variables simultaneously [@bird:2022]. For training readiness, researchers have measured changes in vertical jump height between consecutive training sessions (e.g., [@watkins:2017]). Examining the difference in performance between the two jump variations may also provide insights into the strengths and weaknesses of the athlete to direct future training [@vanHooren2017]. For example, some researchers may compute an "eccentric utilization ratio" by comparing metrics from the countermovement jump relative to the squat jump and use that to inform whether someone should focus their training on improving their ability to leverage the stretch shortening cycle to maximize jumping performance [@vanHooren2017]. Although it is common to collect this kinetic data from a force plate for these jump variations for several applications, there are currently no free, open-source resources to detect events and compute metrics for reproducible and accessible data processing. `JumpMetrics` fills this gap in both applied practice and in the sports science literature.
 
 # Statement of need
 
@@ -61,13 +61,13 @@ Figure 2. Example countermovement jump velocity trace with events detected durin
 
[…]
