### 2026-05-21 f4cd1c2
bumped pixi version for github workflow
