+++ .github/workflows/python-tests.yml
@@ -21,7 +21,7 @@ jobs:
     - name: Install Pixi
       uses: prefix-dev/setup-pixi@v0.8.1
       with:
-        pixi-version: v0.54.2
+        pixi-version: v0.68.1
         cache: true
 
     - name: Install dependencies
