### 2026-08-15 #10651 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
