### 2026-08-31 #10651 @felixchenier
@rkurchin No blocker. I didn't check this checkbox:

Collaborative effort: Does the project show evidence of community engagement — such as contributions from multiple developers, issues/discussions from external users, or community influence on development — in the repo or as described in the paper?

because I don't think there is evidence of it. But it doesn't mean that the paper shouldn't be published.
